# [native] Add a rust implementation of whitespace_parser