--- a/.editorconfig
+++ b/.editorconfig
@@ -1,6 +1,6 @@
 root = true
 
-[*.{py,pyi,toml,md}]
+[*.{py,pyi,rs,toml,md}]
 charset = "utf-8"
 end_of_line = lf
 indent_size = 4
@@ -8,3 +8,7 @@ indent_style = space
 insert_final_newline = true
 trim_trailing_whitespace = true
 max_line_length = 88
+
+[*.rs]
+# https://github.com/rust-dev-tools/fmt-rfcs/blob/master/guide/guide.md
+max_line_length = 100