--- a/setup.py
+++ b/setup.py
@@ -49,6 +49,7 @@ setuptools.setup(
     install_requires=[dep.strip() for dep in open("requirements.txt").readlines()],
     extras_require={
         "dev": [dep.strip() for dep in open("requirements-dev.txt").readlines() if "=" in dep],
+        "native": ["libcst_native==0.1.0"],
     },
     classifiers=[
         "License :: OSI Approved :: MIT License",