@zsol 2021-01-27
Interesting! I wonder how this compares to a `mypyc`-compiled version of the parsing code. I might experiment with that, I had really good results when playing with mypyc in black.

> I haven't done much benchmarking or any profiling yet.

Would be really curious about the bottlenecks and if they're easy to fix.

>  It doesn't currently provide any useful performance advantage, and I can't guarantee that I'll ever get libcst_native this to a state where it provides a useful performance advantage.

Yep, agreed. We can keep the experiment in this PR for now until it provides a significant enough perf advantage to outweigh the maintenance and convenience (to users) cost