@bgw 2021-03-14
Status update!

I've got a working Rust-based tokenizer here: https://github.com/bgw/LibCST/commit/ad0d9006d0e44fc2533e31f3b73b3310c8cc022f

It's based on CPython's C-based tokenizer (not the regex-heavy pure-python parso tokenizer), so it's a little more complicated, but it can also (at least in theory) generate better error messages. It should also handle a few very weird edge cases around formfeed characters (#446) and mixed tab/space indentation better.

Benchmark on master:

```
$ sudo -E perf stat ../libcst-env/bin/python -m libcst.tool print libcst/_nodes/expression.py > /dev/null

 Performance counter stats for '../libcst-env/bin/python -m libcst.tool print libcst/_nodes/expression.py':

          1,919.38 msec task-clock                #    0.971 CPUs utilized
             1,074      context-switches          #    0.560 K/sec
                 1      cpu-migrations            #    0.001 K/sec
            15,501      page-faults               #    0.008 M/sec
     6,664,692,678      cycles                    #    3.472 GHz
     8,909,801,642      instructions              #    1.34  insn per cycle
     1,942,674,148      branches                  # 1012.135 M/sec
        35,986,805      branch-misses             #    1.85% of all branches

       1.977175312 seconds time elapsed

       1.837757000 seconds user
       0.082640000 seconds sys
```

Benchmark on top of https://github.com/bgw/LibCST/commit/ad0d9006d0e44fc2533e31f3b73b3310c8cc022f:

```
sudo -E perf stat ../libcst-env/bin/python -m libcst.tool print libcst/_nodes/expression.py > /dev/null
[sudo] password for bgw:

 Performance counter stats for '../libcst-env/bin/python -m libcst.tool print libcst/_nodes/expression.py':

          1,634.40 msec task-clock                #    0.984 CPUs utilized
               343      context-switches          #    0.210 K/sec
                 0      cpu-migrations            #    0.000 K/sec
            15,426      page-faults               #    0.009 M/sec
     5,695,460,558      cycles                    #    3.485 GHz
     7,868,764,292      instructions              #    1.38  insn per cycle
     1,721,549,089      branches                  # 1053.325 M/sec
        27,874,190      branch-misses             #    1.62% of all branches

       1.661251179 seconds time elapsed

       1.580016000 seconds user
       0.055578000 seconds sys
```

We're getting to the point where there's actually some useful performance gain! However, I think most of the possible performance gains are in porting the CSTNode types, so that's likely my next target. The CST type definitions involve a lot more code than the tokenizer or whitespace parser though, so it'll take even more time.

We currently frequently copy and traverse over CSTNodes. I think we could do a lot better with codegen and a native implementation.