@bgw 2021-03-14
> Interesting! I wonder how this compares to a mypyc-compiled version of the parsing code. I might experiment with that, I had really good results when playing with `mypyc` in black.

Porting all of LibCST to mypyc would be a difficult project. The tokenizer and whitespace parser have type annotations, but the LL(1) parser is nearly impossible to type annotate as it's currently designed. I suppose you could just use mypyc on the parts that are already typechecked.

The CST node types could be compile with mypyc, but the biggest problems with them (unnecessary object clones during traversal) are actually going to need some sort of codegen (or macro) system to solve, so it's not actually a very compiler/language-specific problem.

My real (selfish) motivation behind this PR is that I've fallen in love with Rust, and I want to write more Rust code. If I was still working on LibCST as part of my day job, I'd probably aim for much lower-hanging fruit than rewriting everything in Rust.

> Would be really curious about the bottlenecks and if they're easy to fix.

When we looked at this before, CST node traversal was disproportionately awful. As stated above, we'd need some sort of codegen (or macro system, *cough* Rust proc-macros *cough*) to solve this in any sort of sane way.