@stroxler 2021-12-22
Should we close this?

I believe #566 built off of a version of this work, so it's effectively been merged already?